update CATEGORY for all spending and income

+++ src/role/role.service.ts
@@ -2,10 +2,13 @@ import { BadRequestException, Injectable } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose';
 import { Role } from './schema/role.schema';
+import { AdminService} from '../admin/admin.service';
 
 @Injectable()
 export class RoleService {
-  constructor(@InjectModel(Role.name) private roleModel: Model<Role>) {}
+  constructor(@InjectModel(Role.name) private roleModel: Model<Role>,
+private adminService: AdminService,
+) {}
   async createRoleService(name: string, permissionID: number[]): Promise<Role> {
     const role = await this.roleModel
       .findOne({
@@ -52,12 +55,20 @@ export class RoleService {
   async viewlistRoleService(): Promise<Role[]> {
     return this.roleModel.find().exec();
   }
-  async deleteRoleService(id: string): Promise<{message:string}>{
+async deleteRoleService(id: string): Promise<{message:string}>{
     try {
+      const checkRoleExistInAdmin = await this.adminService.findOneAdminbyIdRoleService(id);
+      if(checkRoleExistInAdmin){
+        throw new BadRequestException('Role exists in admin');
+      }
+      const role = await this.roleModel.findById(id).exec();
+      if (!role) {
+        throw new BadRequestException('Role not exists');
+      }
       await this.roleModel.findByIdAndDelete(id).exec();
       return {message: 'Role deleted successfully'};
     } catch (error) {
-      throw new BadRequestException('Role not exists');
+      throw error;
     }
-  }
+}
 }
